Update to more informed priors
This is based on the following criteria:

* Prior probability of choosing left should be roughly uniform.
* Utility > 4 is rather extreme, as it leads to choice probabilities of >98% (assuming other option is neutral).
* Effect of displaying option on left-hand side should be <= 25 percentage points (which means <= 1).

Before, we had extreme utilities of 10 and more, leading to prior probability of choosing left either almost 0 or almost 1. Effect of displaying option on left-hand side was large (up to 2 on the utility scale, 40 percentage points).

Priors are now more informed and more narrow. My hope is that this helps the sampler, too. On a single test run with the default config, sampling took 93 seconds with the less informed priors and 39 seconds with the more informed priors. Not sure this is generalisable, though.