Reduce number of tuning steps dramatically
According to Statistical Rethinking, only about half of all iterations need to be spent in tuning. Tuning is slow. As long as the sampling stats look fine, tuning steps can be reduced.

I only reduced number of tuning steps for the nocovariates-nocovariances model so far. Likely, this can be done for the other models too.